+++ addons/barcodes/static/tests/basic/legacy/barcode_parser_tests.js
