+++ addons/web/static/src/views/view_button/view_button.js
@@ -129,10 +129,10 @@ ViewButton.props = [
     "id?",
     "tag?",
     "record?",
+    "attrs?",
     "className?",
     "context?",
     "clickParams?",
-    "hotkey?",
     "icon?",
     "defaultRank?",
     "disable?",
